prepareData.kt was refactored

+++ src/main/kotlin/data.kt
@@ -2,7 +2,7 @@ import org.jetbrains.skija.Canvas
 import org.jetbrains.skija.Font
 import org.jetbrains.skija.Paint
 
-class classData(var key : String = "", var values : MutableList<Int> = mutableListOf()){
+class classData(var key : String = "", var values : MutableList<Float> = mutableListOf()){
 
 }
 
@@ -10,7 +10,7 @@ class classGraph(var n : Int = 0, var names : MutableList<String> = mutableListO
 
 }
 
-class classData1(var n : Int = 0, var keys : MutableList<String> = mutableListOf(), var values : MutableList<Int> = mutableListOf()){
+class classData1(var n : Int = 0, var keys : MutableList<String> = mutableListOf(), var values : MutableList<Float> = mutableListOf()){
 
 }
